second try of mainService of graduated
a huge number of corrections of remarks in accordance with the review

+++ stats-service/server/src/main/java/ru/practicum/server/model/ViewStats.java
@@ -14,6 +14,6 @@ import lombok.experimental.FieldDefaults;
 public class ViewStats {
     String app;
     String uri;
-    Long hits;
+    Integer hits;
 
 }
